Rename "ExchangeRatesApiLegacy" driver to "ExchangeRatesApiIo".
